% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{a function which uses the bootstrap method to generate a p-value}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{a vector of data}

\item{conf.level}{the desired confidence level}

\item{iter}{the number of iterations}

\item{mu0}{the value of mu}

\item{test}{the areas to generate}
}
\value{
a list of values
}
\description{
a function which uses the bootstrap method to generate a p-value
}
\examples{
\dontrun{x=rnorm(...);bootpval(x)}
}
