% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsamplezscores.R
\name{getsamplezscores}
\alias{getsamplezscores}
\title{A function which returns a vector of z-scores}
\usage{
getsamplezscores(x)
}
\arguments{
\item{x}{a vector of data}
}
\value{
a vector of z-scores
}
\description{
A function which returns a vector of z-scores
}
\examples{
\dontrun{x <- 1:100; getsamplezscores(x=d)}
}
