% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{A function which generates a binomial distribution a given number of times}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{the desired number of iterations}

\item{n}{number of observations}

\item{p}{probability}
}
\value{
a table with n entries containing probabilities
}
\description{
A function which generates a binomial distribution a given number of times
}
\examples{
\dontrun{mybin(iter=100, n=10, p=0.5)}
}
