% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{a function which creates a confidence interval and a point estimate given a sample}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{function to estimate}

\item{alpha}{1-alpha confidence interval}

\item{cx}{text scaling}

\item{...}{graphical parameters}
}
\value{
a list of parameters
}
\description{
a function which creates a confidence interval and a point estimate given a sample
}
\examples{
\dontrun{sam=sample(...); myboot2(x=sam,fun="mean")}
}
