% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{A function which creates a confidence interval for the mean}
\usage{
myci(x)
}
\arguments{
\item{x}{a vector of data}
}
\value{
a confidence interval
}
\description{
A function which creates a confidence interval for the mean
}
\examples{
\dontrun{sam=qnorm(30,mean=10,sd=12);myci(sam)}
}
