% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfunction.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{First Function}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{A vector of quantitative data}
}
\value{
A vector of squared components
}
\description{
First Function
}
\examples{
y <- 1:10; myfirstfunction(y)
}
