% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{a function which creates a maximum likelihood estimate}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{a function for maximum likelihood}

\item{x}{a vector of data}

\item{param}{a sequence with a given length}

\item{...}{graphical parameters}
}
\value{
a list of values and a plot
}
\description{
a function which creates a maximum likelihood estimate
}
\examples{
\dontrun{func=function(x) ...;mymaxlik(lfun=func, x, param=seq(0,10, length=1000), main="Title")}
}
