% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{a function which creates a curve and finds the density of a given lower tail probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard devation}

\item{a}{the value at which the find a lower tail probability}
}
\value{
a list containing the area
}
\description{
a function which creates a curve and finds the density of a given lower tail probability
}
\examples{
\dontrun{getsamplezscores(mu=5, sigma=1, a=5)}
}
